Initial Commit
I'm curious about recording audio in the browser, and how one could maybe make a DAW or something, maybe it's not very feasible though. Even like an Electron/Tauri app or something in Node with the CLI could be kind of interesting. How do native apps deal with audio? If it's backed by a native implementation, couldn't it be something web-based like this as well?

https://www.reddit.com/r/linuxaudio/comments/omchul/focusrite_scarlett_2i2_3rd_gen_appearing_as_a/
https://stackoverflow.com/questions/53218829/node-js-access-audio-interface
https://developer.mozilla.org/en-US/docs/Web/API/MediaStream_Recording_API/Using_the_MediaStream_Recording_API
https://developer.mozilla.org/en-US/docs/Web/API/MediaRecorder
https://ralzohairi.medium.com/audio-recording-in-javascript-96eed45b75ee ⭐️
https://stackoverflow.com/questions/63521599/setting-audio-element-source-from-blob
https://franzeus.medium.com/record-audio-in-js-and-upload-as-wav-or-mp3-file-to-your-backend-1a2f35dea7e8